+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 #### **resource-metadata-sqs**
 ### 🐛 Bug Fix 🔧
 - Add missing env var `EC2_CHUNK_SIZE`
-- Add inline policy to assume cross-config AWS Config role
+- Add inline policy to assume cross-account AWS Config role
 
 ## v3.19.3
 #### **resource-metadata-sqs**